pyproject.toml
[project]
name = "support101"
version = "1.0.0"
description = "AI-powered customer support platform with RAG, agent framework, and continuous learning"
requires-python = ">=3.11"
dependencies = [
    # Core Framework
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.10.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    # Database
    "SQLAlchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "alembic>=1.14.0",
    # Auth
    "PyJWT>=2.8.0",
    "python-jose>=3.3.0",
    # Caching & Rate Limiting
    "redis>=5.0.3",
    "fastapi-cache2>=0.2.1",
    "fastapi-limiter>=0.1.5",
    # LLM & RAG
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "pinecone>=5.0.0",
    "fastembed>=0.4.0",
    "sentence-transformers>=3.0.0",
    "huggingface-hub>=0.26.0",
    "tenacity>=9.0.0",
    # Observability
    "prometheus_client>=0.21.0",
    # Document Processing
    "pdfplumber>=0.11.0",
    # Scheduling
    "schedule>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "pendulum>=3.0.0",
    "mutmut",
    "ruff>=0.8.0",
]
anthropic = ["langchain-anthropic>=0.3.0"]
google = ["langchain-google-genai>=2.0.0"]
tracing = [
    "traceloop-sdk>=0.30.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp-proto-http>=1.28.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]
ignore = ["E501", "B008"]

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = ["B017"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "packages"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
